requirements.txt
line-bot-sdk==3.5.0
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
firebase-admin==6.2.0
python-dateutil==2.8.2 